chore: logger use from param & type compile
