+++ .changeset/tough-pants-switch.md
@@ -1,0 +1,6 @@
+---
+"@farmfe/core": patch
+---
+
+- core types should be included in the compilation process
+- logger used should be obtained from the parameter
