+++ cspell.json
@@ -58,9 +58,12 @@
     "envinfo",
     "esbuild",
     "execa",
+    "fargs",
     "farmfe",
     "flac",
+    "fnames",
     "Fock",
+    "funs",
     "gnueabihf",
     "guolao",
     "hashbrown",
@@ -68,6 +71,7 @@
     "icns",
     "idents",
     "IHDR",
+    "iife",
     "importee",
     "Inctive",
     "indicatif",
@@ -158,6 +162,7 @@
     "tsup",
     "Turbopack",
     "Typecheck",
+    "typeofs",
     "unocss",
     "unplugin",
     "unref",
