+++ packages/core/src/config/normalize-config/normalize-output.ts
@@ -1,6 +1,6 @@
 import { browsersWithSupportForFeatures } from 'browserslist-generator';
 
-import { Config } from '../../../binding/index.js';
+import { Config } from '../../types/binding.js';
 import {
   FARM_TARGET_BROWSER_ENVS,
   mapTargetEnvValue
