+++ packages/core/src/plugin/type.ts
@@ -1,3 +1,4 @@
+import { Compiler, ResolvedUserConfig, Server, UserConfig } from '../index.js';
 import {
   Config,
   PluginLoadHookParam,
@@ -6,8 +7,7 @@ import {
   PluginResolveHookResult,
   PluginTransformHookParam,
   PluginTransformHookResult
-} from '../../binding/index.js';
-import { Compiler, ResolvedUserConfig, Server, UserConfig } from '../index.js';
+} from '../types/binding.js';
 
 // https://stackoverflow.com/questions/61047551/typescript-union-of-string-and-string-literals
 // eslint-disable-next-line @typescript-eslint/ban-types
@@ -125,7 +125,7 @@ type Callback<P, R> = (
   param: P,
   context?: CompilationContext,
   hookContext?: { caller?: string; meta: Record<string, unknown> }
-) => Promise<R | null | undefined>;
+) => Promise<R | null | undefined> | R | null | undefined;
 type JsPluginHook<F, P, R> = { filters: F; executor: Callback<P, R> };
 
 export interface JsPlugin {
@@ -245,4 +245,4 @@ export type {
   PluginLoadHookResult,
   PluginTransformHookParam,
   PluginTransformHookResult
-} from '../../binding/index.js';
+} from '../types/binding.js';
