+++ packages/core/src/utils/share.ts
@@ -4,7 +4,7 @@ import os from 'node:os';
 import path, { dirname } from 'node:path';
 import readline from 'node:readline';
 import { fileURLToPath } from 'node:url';
-import { Config } from '../../binding/index.js';
+import { Config } from '../types/binding.js';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore import packageJson from '../../package.json';
